Moves SnapshotGossipManager to its own file
